Мониторинг задач редмайна

+++ src/main/java/ru/arriah/redminenotification/redmine/task/Schedulable.java
@@ -1,5 +1,5 @@
 package ru.arriah.redminenotification.redmine.task;
 
 interface Schedulable extends Runnable {
-   int getDelay();
+   long getDelay();
 }
